Split out sandbox capability-creating logic.
